+++ src/BusinessLogic/Resources/js/PickShippingServiceController.js
@@ -10,6 +10,8 @@ if (!window.Packlink) {
      * @property {string} getTaskStatusUrl
      * @property {string} startAutoConfigureUrl
      * @property {string} disableCarriersUrl
+     * @property {string} getCurrencyDetails
+     * @property {string} systemId
      * @property {boolean} newService
      */
 
@@ -38,6 +40,11 @@ if (!window.Packlink) {
         let noServicesModal;
 
         /**
+         * @type string
+         */
+        let defaultCurrency;
+
+        /**
          * Displays page content.
          *
          *  @param {{from: string, newService: boolean}} config
@@ -45,6 +52,7 @@ if (!window.Packlink) {
         this.display = function (config) {
             utilityService.showSpinner();
             templateService.setCurrentTemplate(templateId);
+            ajaxService.get(configuration.getCurrencyDetails, getDefaultCurrency);
             ajaxService.get(configuration.getTaskStatusUrl, (response) => {
                 checkServicesStatus(response, config);
             });
@@ -151,6 +159,25 @@ if (!window.Packlink) {
         };
 
         /**
+         * Retrieves system info.
+         *
+         * @param {SystemInfo[]} systemInfos
+         */
+        const getDefaultCurrency = (systemInfos) => {
+            let systemInfo = systemInfos[0];
+
+            if (configuration.systemId !== null) {
+                systemInfos.forEach((info) => {
+                    if (info.systemId === configuration.systemId) {
+                        systemInfo = info;
+                    }
+                });
+            }
+
+            defaultCurrency = systemInfo.currencies[0];
+        };
+
+        /**
          * Binds services.
          *
          * @param {ShippingService[]} services
@@ -196,7 +223,7 @@ if (!window.Packlink) {
                 table = templateService.getComponent('pl-shipping-services-table'),
                 list = templateService.getComponent('pl-shipping-services-list').querySelector('.pl-shipping-services-list'),
                 render = (elem, id, tag) => {
-                    Packlink.ShippingServicesRenderer.render(elem, id, tag, filteredServices, false, handleServiceAction);
+                    Packlink.ShippingServicesRenderer.render(elem, id, tag, filteredServices, false, handleServiceAction, defaultCurrency);
                 };
 
             render(table.querySelector('tbody'), 'pl-shipping-services-row', 'tr');