+++ src/DemoUI/src/Bootstrap.php
@@ -36,6 +36,8 @@ use Packlink\DemoUI\Repository\SessionRepository;
 use Packlink\DemoUI\Services\BusinessLogic\CarrierService;
 use Packlink\DemoUI\Services\BusinessLogic\ConfigurationService;
 use Packlink\DemoUI\Services\BusinessLogic\ShopOrderService;
+use Packlink\BusinessLogic\SystemInformation\SystemInfoService as SystemInfoServiceInterface;
+use Packlink\DemoUI\Services\BusinessLogic\SystemInfoService;
 use Packlink\DemoUI\Services\Infrastructure\LoggerService;
 
 /**
@@ -83,6 +85,10 @@ class Bootstrap extends BootstrapComponent
      * @var RegistrationInfoService
      */
     private $registrationInfoService;
+    /**
+     * @var SystemInfoService
+     */
+    private $systemInfoService;
 
     /**
      * Bootstrap constructor.
@@ -97,6 +103,7 @@ class Bootstrap extends BootstrapComponent
         $this->carrierService = new CarrierService();
         $this->userAccountService = UserAccountService::getInstance();
         $this->registrationInfoService = new TestRegistrationInfoService();
+        $this->systemInfoService = new SystemInfoService();
     }
 
     /**
@@ -200,6 +207,13 @@ class Bootstrap extends BootstrapComponent
             RegistrationInfoService::CLASS_NAME,
             function () use ($instance) {
                 return $instance->registrationInfoService;
+            }
+        );
+
+        ServiceRegister::registerService(
+            SystemInfoServiceInterface::CLASS_NAME,
+            function () use ($instance) {
+                return $instance->systemInfoService;
             }
         );
     }